+++ src/pages/Main/TodoList/style.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const ToDoListWrapper = styled.div`
+  width: 100%;
+  height: 100%;
+`;
